Story #138 (Renu/Yue) Test to save the Bill details

+++ src/test/java/com/thoughtworks/twu/domain/BillTest.java
@@ -2,6 +2,8 @@ package com.thoughtworks.twu.domain;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
@@ -9,9 +11,9 @@ public class BillTest {
 
     @Test
     public void shouldHaveDescription() throws Exception {
-        Bill bill = new Bill("Dinner",2000);
+        Bill bill = new Bill("Dinner",2000.0);
         String expectedDescription = "Dinner";
-        int expectedAmount = 2000;
+        double expectedAmount = 2000.0;
         assertThat(bill.getDescription(),is(expectedDescription));
         assertThat(bill.getAmount(),is(expectedAmount));
     }
